Move org.sonatype.nexus.util.sequence.* to org.sonatype.nexus.proxy.utils
