+++ components/nexus-core/src/main/java/org/sonatype/nexus/proxy/walker/FixedRateWalkerThrottleController.java
@@ -14,7 +14,7 @@ package org.sonatype.nexus.proxy.walker;
 
 import java.util.concurrent.TimeUnit;
 
-import org.sonatype.nexus.util.sequence.NumberSequence;
+import org.sonatype.nexus.proxy.utils.NumberSequence;
 
 import com.google.common.base.Preconditions;
