+++ components/nexus-core/src/test/java/org/sonatype/nexus/proxy/utils/NumberSequenceTest.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.util.sequence;
+package org.sonatype.nexus.proxy.utils;
 
 import org.junit.Assert;
 import org.junit.Test;
