+++ components/nexus-core/src/test/java/org/sonatype/nexus/proxy/walker/FixedRateWalkerThrottleControllerTest.java
@@ -12,10 +12,10 @@
  */
 package org.sonatype.nexus.proxy.walker;
 
+import org.sonatype.nexus.proxy.utils.ConstantNumberSequence;
+import org.sonatype.nexus.proxy.utils.FibonacciNumberSequence;
+import org.sonatype.nexus.proxy.utils.NumberSequence;
 import org.sonatype.nexus.proxy.walker.WalkerThrottleController.ThrottleInfo;
-import org.sonatype.nexus.util.sequence.ConstantNumberSequence;
-import org.sonatype.nexus.util.sequence.FibonacciNumberSequence;
-import org.sonatype.nexus.util.sequence.NumberSequence;
 import org.sonatype.sisu.litmus.testsupport.TestSupport;
 
 import org.junit.Test;
